check_answer function created
upon answering in the entry box, the answer gets validated using check_answer() function i added. the input is checked to be a valid number and then checked to be correct or not. if the answer is correct the first time 10 points are added, if the answer is right the second try then only 5 points are added. If the answer is wrong both times then no points are added and the correct answer gets displayed